requirements.txt
discord.py
torch
torchvision
Pillow
requests
Flask
python-dotenv
git+https://github.com/openai/CLIP.git